+++ SegPL_U.py
@@ -167,7 +167,7 @@ class SegPL_U:
                     else:   
                         probabilities = torch.nn.Softmax(dim=1)(logits_x_ulb)
                         
-                    pseudo_labels = (probabilities>p_cutoff).long().detach()                    
+                    pseudo_labels = (probabilities>p_cutoff).float().detach()                    
                     
                     
                     mask_pl = (probabilities>threshold).long() + (probabilities<1 - threshold).long()
@@ -182,7 +182,7 @@ class SegPL_U:
                     anti_pseudo_labels = (probabilities>p_cutoff).long().detach()
                     
                     
-                    mask_anti_pl = (probabilities>threshold).long() + (probabilities<1 - threshold).long()
+                    mask_anti_pl = (probabilities>threshold).float() + (probabilities<1 - threshold).long()
                     anti_unsup_loss = self.unsupervised_loss(logits_x_lb, anti_pseudo_labels, mask=mask_anti_pl[:,:1])
                     
                     if args.debiased:
